app/requirements.txt
Django==4.2.2
djangorestframework==3.14.0
psycopg2-binary==2.9
gunicorn==20.1.0
Pillow==9.5.0
drf-spectacular==0.26.2
drf-spectacular-sidecar==2023.6.1
django-filter==23.2
djangorestframework-simplejwt==5.2.2